# Conflicts:
#	Caravela.Framework.Impl/Linking/LinkerRewritingDriver.Events.cs
#	Caravela.Framework.Impl/Linking/LinkerRewritingDriver.Methods.cs
#	Caravela.Framework.Impl/Linking/LinkerRewritingDriver.Properties.cs
#	Caravela.Framework.Impl/Linking/LinkerRewritingDriver.cs

+++ Tests/Caravela.Framework.Tests.InternalPipeline/Tests/Templating/NamespaceExpansion/PartiallyQualified.t.cs
@@ -1,5 +1,5 @@
 int Method(int a)
 {
     var c = new global::Caravela.Framework.Tests.Integration.TestInputs.Templating.NamespaceExpansion.PartiallyQualified.ChildNs.ChildClass();
-    return (int)this.Method(a);
+    return this.Method(a);
 }
