+++ Tests/Caravela.Framework.Tests.InternalPipeline/Tests/Templating/Syntax/Lambdas/CompileTimeLinqSimpleLambda.t.cs
@@ -1,5 +1,5 @@
 int Method(int a, int b)
 {
     global::System.Console.WriteLine(0);
-    return (int)this.Method(a, b);
+    return this.Method(a, b);
 }
