+++ Tests/Caravela.Framework.Tests.InternalPipeline/Tests/Templating/Syntax/New/NewClassCompileTime.t.cs
@@ -2,5 +2,5 @@ int Method(int a)
 {
     global::System.Console.WriteLine("string");
     global::System.Console.WriteLine("string");
-    return (int)this.Method(a);
+    return this.Method(a);
 }
